requirements.txt
flask==3.1.2
pandas==3.0.0
scikit-learn==1.8.0
requests==2.32.5
pyarrow==23.0.0
weasyprint==68.1
joblib==1.5.3
python-dotenv==1.2.1
